Move dataset feature migration tests to its endpoint test file
